refactor(riversctl): split main.rs into 6 command modules
Move 913-LOC monolith into commands/{start,doctor,validate,exec,admin}.rs.
main.rs retains entry point, dispatch match, print_usage, and tests (210 LOC).
