+++ crates/riversctl/src/commands/mod.rs
@@ -1,0 +1,5 @@
+pub mod start;
+pub mod doctor;
+pub mod validate;
+pub mod exec;
+pub mod admin;
